requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
numpy>=1.19.5
Pillow>=8.3.1
scikit-learn>=0.24.2
matplotlib>=3.4.2
wandb>=0.12.0
PyYAML>=5.4.1
tqdm>=4.61.2
einops>=0.3.2

